MAINT: Move add_newdocs into core, since it only adds docs to those pieces
